requirements.txt
alembic==1.18.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
ecdsa==0.19.1
fastapi==0.128.0
greenlet==3.3.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.1
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pgvector==0.4.2
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
redis==7.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.14
